+++ library/core/src/test/java/com/google/android/exoplayer2/CTest.java
@@ -22,13 +22,11 @@ import android.media.MediaCodec;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
-import org.robolectric.annotation.Config;
 
 /**
  * Unit test for {@link C}.
  */
 @RunWith(RobolectricTestRunner.class)
-@Config(sdk = Config.TARGET_SDK, manifest = Config.NONE)
 public class CTest {
 
   @SuppressLint("InlinedApi")