+++ library/core/src/test/java/com/google/android/exoplayer2/extractor/ExtractorTest.java
@@ -21,13 +21,11 @@ import com.google.android.exoplayer2.C;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
-import org.robolectric.annotation.Config;
 
 /**
  * Unit test for {@link Extractor}.
  */
 @RunWith(RobolectricTestRunner.class)
-@Config(sdk = Config.TARGET_SDK, manifest = Config.NONE)
 public final class ExtractorTest {
 
   @Test