+++ library/core/src/test/java/com/google/android/exoplayer2/offline/ActionFileTest.java
@@ -31,13 +31,11 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
 import org.robolectric.RuntimeEnvironment;
-import org.robolectric.annotation.Config;
 
 /**
  * Unit tests for {@link ProgressiveDownloadAction}.
  */
 @RunWith(RobolectricTestRunner.class)
-@Config(sdk = Config.TARGET_SDK, manifest = Config.NONE)
 public class ActionFileTest {
 
   private File tempFile;